% Generated by roxygen2 (4.0.2): do not edit by hand
\name{append_to}
\alias{append_to}
\title{assumes that tbl is an ffdf and that dat has same structure as tbl}
\usage{
append_to(tbl, dat, check_structure = TRUE, ...)
}
\description{
assumes that tbl is an ffdf and that dat has same structure as tbl
}
\keyword{internal}

