% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as_ffdf}
\alias{as_ffdf}
\title{Convert a data.frame to a ffdf object}
\usage{
as_ffdf(x, character_to_factor = TRUE, ...)
}
\arguments{
\item{x}{a data.frame or coercable to a data.frame}

\item{character_to_factor}{\code{logical} will coerce a character to factor,
since those types cannot be stored in \code{ff}.}
}
\description{
Improved version of as.ffdf:
- Automatically converts characters to factors: ff does not support character
  vectors
- allow for zero row data.frames
}

