% Generated by roxygen2 (4.0.1): do not edit by hand
\name{grouped_ffdf}
\alias{grouped_ffdf}
\alias{is.grouped_ffdf}
\title{A grouped ffdf.}
\usage{
grouped_ffdf(data, vars, drop = TRUE)

is.grouped_ffdf(x)
}
\arguments{
\item{data}{a data source or data frame.}

\item{vars}{a list of quoted variables.}

\item{drop}{}

\item{x}{an object to check}
}
\description{
The easiest way to create a grouped ffdf is to call the \code{group_by}
method on a ffdf or data source: this will take care of capturing
the unevalated expressions for you.
}

