% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rename_.grouped_ffdf}
\alias{arrange_.ffdf}
\alias{arrange_.tbl_ffdf}
\alias{do.ffdf}
\alias{do.tbl_ffdf}
\alias{filter_.ffdf}
\alias{filter_.tbl_ffdf}
\alias{manip_ffdf}
\alias{mutate_.ffdf}
\alias{mutate_.tbl_ffdf}
\alias{rename_.ffdf}
\alias{rename_.grouped_ffdf}
\alias{rename_.tbl_ffdf}
\alias{select_.ffdf}
\alias{select_.tbl_ffdf}
\alias{slice_.ffdf}
\alias{slice_.tbl_ffdf}
\alias{summarise_.ffdf}
\alias{summarise_.tbl_ffdf}
\title{Data manipulation for ffdf.}
\usage{
\method{rename_}{grouped_ffdf}(.data, ..., .dots)

\method{filter_}{ffdf}(.data, ..., .dots)

\method{filter_}{tbl_ffdf}(.data, ..., .dots)

\method{summarise_}{ffdf}(.data, ..., .dots)

\method{summarise_}{tbl_ffdf}(.data, ...)

\method{mutate_}{ffdf}(.data, ..., .dots, inplace = FALSE)

\method{mutate_}{tbl_ffdf}(.data, ...)

\method{arrange_}{ffdf}(.data, ..., .dots)

\method{arrange_}{tbl_ffdf}(.data, ...)

\method{select_}{ffdf}(.data, ..., .dots)

\method{select_}{tbl_ffdf}(.data, ..., .dots)

\method{rename_}{ffdf}(.data, ..., .dots)

\method{rename_}{tbl_ffdf}(.data, ..., .dots)

\method{slice_}{ffdf}(.data, ..., .dots)

\method{slice_}{tbl_ffdf}(.data, ..., .dots)

\method{do}{ffdf}(.data, ...)

\method{do}{tbl_ffdf}(.data, ...)
}
\arguments{
\item{.data}{a ffdf}

\item{...}{variables interpreted in the context of \code{.data}}

\item{inplace}{if \code{FALSE} (the default) the data frame will be copied
prior to modification to avoid changes propagating via reference.}

\item{.env}{The environment in which to evaluate arguments not included in
the data. The default should suffice for ordinary usage.}
}
\description{
Data manipulation for ffdf.
}
\examples{
if (require("hflights")) {
hflights2 <- tbl_ffdf(hflights)
filter(hflights2, Month == 1, DayofMonth == 1, Dest == "DFW")
head(select(hflights2, Year:DayOfWeek))
summarise(hflights2, delay = mean(ArrDelay, na.rm = TRUE), n = length(ArrDelay))
head(mutate(hflights2, gained = ArrDelay - DepDelay))
head(arrange(hflights2, Dest, desc(ArrDelay)))
}
}

