% Generated by roxygen2 (4.0.1): do not edit by hand
\name{manip_grouped_ffdf}
\alias{arrange.grouped_ffdf}
\alias{filter.grouped_ffdf}
\alias{manip_grouped_ffdf}
\alias{mutate.grouped_ffdf}
\alias{summarise.grouped_ffdf}
\title{Data manipulation for grouped data tables.}
\usage{
\method{filter}{grouped_ffdf}(.data, ..., env = parent.frame())

\method{summarise}{grouped_ffdf}(.data, ...)

\method{mutate}{grouped_ffdf}(.data, ..., inplace = FALSE)

\method{arrange}{grouped_ffdf}(.data, ...)
}
\arguments{
\item{.data}{a data table}

\item{...}{variables interpreted in the context of \code{.data}}

\item{inplace}{if \code{FALSE} (the default) the data frame will be copied
prior to modification to avoid changes propagating via reference.}
}
\description{
Data manipulation for grouped data tables.
}
\examples{
data("baseball", package = "plyr")
baseball_ffdf <- as.ffdf(baseball)
players <- group_by(baseball_ffdf, id)
filter(players, g == max(g))
summarise(players, g = mean(g))
mutate(players, year = year - min(year) + 1)
arrange(players, id, desc(year))
select(players, id:team)

# All manip functions preserve grouping structure, except for summarise
# (for hopefully obvious reasons)
by_year <- mutate(players, cyear = year - min(year) + 1)
summarise(by_year, years = max(cyear))

# You can also manually ungroup:
arrange(ungroup(by_year), id, year)
}

