% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sample_data.default}
\alias{sample_data.default}
\title{Uniformly sample data tabular object}
\usage{
sample_data.default(x, n = 10000, seed = NULL, ...)
}
\arguments{
\item{x}{tabular object to sample from}

\item{n}{size of sample}

\item{seed}{random seed}

\item{...}{}
}
\description{
Uniformly sample data tabular object
}

