% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sample_data}
\alias{sample_data}
\title{Uniformly sample data tabular object}
\usage{
sample_data(x, n = 10000, seed = NULL, ...)
}
\arguments{
\item{x}{tabular object to sample from}

\item{n}{size of sample}

\item{seed}{random seed}

\item{...}{not used}
}
\description{
Uniformly sample data tabular object
}

