% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tbl_ffdf}
\alias{as.data.frame.tbl_ffdf}
\alias{as.tbl.ffdf}
\alias{head.tbl_ffdf}
\alias{print.tbl_ffdf}
\alias{tail.tbl_ffdf}
\alias{tbl_ffdf}
\alias{tbl_vars.ffdf}
\alias{tbl_vars.tbl_ffdf}
\title{Create a ffdf tbl object}
\usage{
tbl_ffdf(data, src = NULL, from = deparse(substitute(data)), ...)

\method{as.tbl}{ffdf}(x, ...)

\method{tbl_vars}{tbl_ffdf}(x)

\method{tbl_vars}{ffdf}(x)

\method{as.data.frame}{tbl_ffdf}(x, row.names = NULL, optional = FALSE, ...)

\method{print}{tbl_ffdf}(x, ..., n = NULL)

\method{head}{tbl_ffdf}(x, n = 6L, ...)

\method{tail}{tbl_ffdf}(x, n = 6L, ...)
}
\arguments{
\item{data}{a ffdf data.frame, will be converted to ffdf using as.ffdf}

\item{...}{additional arguments to be passed to or from methods.}

\item{x}{any \R object.}

\item{row.names}{\code{NULL} or a character vector giving the row
    names for the data frame.  Missing values are not allowed.}

\item{optional}{logical. If \code{TRUE}, setting row names and
    converting column names (to syntactic names: see
    \code{\link{make.names}}) is optional.}
}
\description{
A ffdf source wraps a local ffdf.
}
\examples{
ds <- tbl_ffdf(mtcars)
ds
}

