% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/append_to.R
\name{append_to}
\alias{append_to}
\title{assumes that tbl is an ffdf and that dat has same structure as tbl}
\usage{
append_to(tbl, dat, check_structure = TRUE, ...)
}
\description{
assumes that tbl is an ffdf and that dat has same structure as tbl
}
\keyword{internal}

