% Generated by roxygen2 (4.0.1): do not edit by hand
\name{manip_ffdf}
\alias{arrange.ffdf}
\alias{arrange.tbl_ffdf}
\alias{do.ffdf}
\alias{do.tbl_ffdf}
\alias{filter.ffdf}
\alias{filter.tbl_ffdf}
\alias{manip_ffdf}
\alias{mutate.ffdf}
\alias{mutate.tbl_ffdf}
\alias{rename.ffdf}
\alias{select.ffdf}
\alias{select.tbl_ffdf}
\alias{summarise.ffdf}
\alias{summarise.tbl_ffdf}
\title{Data manipulation for ffdf.}
\usage{
\method{filter}{ffdf}(.data, ..., .env = parent.frame())

\method{filter}{tbl_ffdf}(.data, ..., .env = parent.frame())

\method{summarise}{ffdf}(.data, ...)

\method{summarise}{tbl_ffdf}(.data, ...)

\method{mutate}{ffdf}(.data, ..., inplace = FALSE)

\method{mutate}{tbl_ffdf}(.data, ...)

\method{arrange}{ffdf}(.data, ...)

\method{arrange}{tbl_ffdf}(.data, ...)

\method{select}{ffdf}(.data, ...)

\method{select}{tbl_ffdf}(.data, ...)

rename.ffdf(.data, ...)

\method{do}{ffdf}(.data, ...)

\method{do}{tbl_ffdf}(.data, ...)
}
\arguments{
\item{.data}{a ffdf}

\item{...}{variables interpreted in the context of \code{.data}}

\item{inplace}{if \code{FALSE} (the default) the data frame will be copied
prior to modification to avoid changes propagating via reference.}

\item{.env}{The environment in which to evaluate arguments not included in
the data. The default should suffice for ordinary usage.}
}
\description{
Data manipulation for ffdf.
}
\examples{
if (require("hflights")) {
hflights2 <- tbl_ffdf(hflights)
filter(hflights2, Month == 1, DayofMonth == 1, Dest == "DFW")
head(select(hflights2, Year:DayOfWeek))
summarise(hflights2, delay = mean(ArrDelay, na.rm = TRUE), n = length(ArrDelay))
head(mutate(hflights2, gained = ArrDelay - DepDelay))
head(arrange(hflights2, Dest, desc(ArrDelay)))
}
}

