% Generated by roxygen2 (4.0.2): do not edit by hand
\name{manip_grouped_ffdf}
\alias{arrange_.grouped_ffdf}
\alias{filter_.grouped_ffdf}
\alias{manip_grouped_ffdf}
\alias{mutate.grouped_ffdf}
\alias{slice_.grouped_ffdf}
\alias{summarise_.grouped_ffdf}
\title{Data manipulation for grouped data tables.}
\usage{
\method{filter_}{grouped_ffdf}(.data, ..., .dots)

\method{summarise_}{grouped_ffdf}(.data, ..., .dots)

mutate.grouped_ffdf(.data, ..., inplace = FALSE)

\method{arrange_}{grouped_ffdf}(.data, ..., .dots)

\method{slice_}{grouped_ffdf}(.data, ..., .dots)
}
\arguments{
\item{.data}{a data table}

\item{...}{variables interpreted in the context of \code{.data}}

\item{inplace}{if \code{FALSE} (the default) the data frame will be copied
prior to modification to avoid changes propagating via reference.}
}
\description{
Data manipulation for grouped data tables.
}
\examples{
species <-
  iris \%>\% tbl_ffdf \%>\%
  group_by(Species)
filter(species, Petal.Width == max(Petal.Width))
summarise(species, Petal.Width=mean(Petal.Width))
}

