% Generated by roxygen2 (4.0.2): do not edit by hand
\name{manip_grouped_ffdf}
\alias{arrange.grouped_ffdf}
\alias{filter.grouped_ffdf}
\alias{manip_grouped_ffdf}
\alias{mutate.grouped_ffdf}
\alias{summarise.grouped_ffdf}
\title{Data manipulation for grouped data tables.}
\usage{
filter.grouped_ffdf(.data, ...)

summarise.grouped_ffdf(.data, ...)

mutate.grouped_ffdf(.data, ..., inplace = FALSE)

arrange.grouped_ffdf(.data, ...)
}
\arguments{
\item{.data}{a data table}

\item{...}{variables interpreted in the context of \code{.data}}

\item{inplace}{if \code{FALSE} (the default) the data frame will be copied
prior to modification to avoid changes propagating via reference.}
}
\description{
Data manipulation for grouped data tables.
}
\examples{
species <-
  iris \%>\% tbl_ffdf \%>\%
  group_by(Species)
filter(species, Petal.Width == max(Petal.Width))
summarise(species, Petal.Width=mean(Petal.Width))
}

