% Generated by roxygen2 (4.0.1): do not edit by hand
\name{src_ffdf}
\alias{src_ffdf}
\title{Open or create ffdf source}
\usage{
src_ffdf(path, ...)
}
\arguments{
\item{path}{\code{character} directory path}

\item{...}{not used}
}
\description{
An ffdf source is a directory in which ffdf tables will reside.
}

